requirements.txt
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
streamlit==1.35.0
matplotlib==3.8.4
